refactor: extract Netch.Utils.Subscription
